// src/app/datos.interfaces.ts
export interface Formulario {
    id: any;
    userId: string;
    nombreCompleto: string;
    tipoDocumento: string;
    numeroDocumento: string;
    direccion: string;
    barrio: string;
    celular: string;
    correo: string;
    seguroSocial: string;
    numeroPersonasHogar: number;
    numeroMayoresEdad: number;
    numeroMenoresEdad: number;
    numeroFormacionDeportiva: number;
    nombreIntegra1: string;
    EdadIntegrante1: string;
    EscolarizadoIntegra1: string;
    Integrante1Disciplina: string;
    nombreIntegra2: string;
    EdadIntegrante2: string;
    EscolarizadoIntegra2: string;
    Integrante2Disciplina: string;
    nombreIntegra3: string;
    EdadIntegrante3: string;
    EscolarizadoIntegra3: string;
    Integrante3Disciplina: string;
    nombreIntegra4: string;
    EdadIntegrante4: string;
    EscolarizadoIntegra4: string;
    Integrante4Disciplina: string;
    ProblemaBarrio: string;
    OtroProBarrio: string;
    ProblemaSoledad: string;
    InformadoCampañas: string;
  }
  